+++ inst/paramtest/test_paramtest_earth_regr_earth.R
@@ -3,18 +3,19 @@ install_learners("regr.earth")
 
 test_that("regr.earth", {
   learner = lrn("regr.earth")
-  fun = earth::earth
+  fun = earth:::earth.default
   exclude = c(
-
+    "x", # handled internally
+    "y", # handled internally
+    "subset", # handled internally
+    "weights", # handled internally
+    "na.action", # only one option
+    "glm" # handled internally
   )
 
   ParamTest = run_paramtest(learner, fun, exclude)
-  expect_true(ParamTest, info = paste0(
-    "
-Missing parameters:
-",
-    paste0("- '", ParamTest$missing, "'", collapse = "
-")))
+  expect_true(ParamTest, info = paste0("Missing parameters:",
+    paste0("- '", ParamTest$missing, "'", collapse = "")))
 })
 
 test_that("regr.earth predict", {
@@ -24,7 +25,6 @@ test_that("regr.earth predict", {
     "object", # handled internally
     "newdata", # handled internally
     "type", # handled internally
-    "interval", # unused
     "level" # unused
   )